Refactor: rename `condlictModals.jsx` -> `conflictModals.jsx`
