+++ frontend/src/components/NewProject/Upload/index.jsx
@@ -10,7 +10,7 @@ import { createRepo, repoExists } from '@utils/giteaApi'
 import { slugifiedNameFromFiles } from '@utils/index'
 import DropZone from '@components/DropZone'
 import { UploadOp, NoOp } from '../Ops'
-import { UploadConflictModal } from '../CondlictModals'
+import { UploadConflictModal } from '../ConflictModals'
 import styles from './index.module.scss'
 
 const Upload = ({ setUserOp }) => {
